+++ test/unit/tools/cane.rb
@@ -4,8 +4,9 @@ module Test
       # Test for the 'cane' tool within the quality gem
       module Cane
         def cane_expected_args
-          "-f '{Rakefile,*.{rb,rake},{src,app,lib,test,spec,feature}/**/" \
-          "*.{rb,rake}}'"
+          "-f '{Rakefile,*.{rb,rake,gemspec}," \
+          '{src,app,lib,test,spec,feature}/**/' \
+          "*.{rb,rake,gemspec}}'"
         end
 
         def expect_cane_run(quality_checker)