+++ test/unit/tools/punchlist.rb
@@ -4,9 +4,10 @@ module Test
       # Test for the 'bigfiles' tool within the quality gem
       module Punchlist
         def punchlist_expected_args
-          "--glob '{Rakefile,Dockerfile,*.{rb,rake,swift,cpp,c,java,py," \
-          'clj,cljs,scala,js,yml,sh,json},{src,app,lib,test,spec,feature}' \
-          "/**/*.{rb,rake,swift,cpp,c,java,py,clj,cljs,scala,js,yml,sh,json}}'"
+          "--glob '{Rakefile,Dockerfile,*.{rb,rake,gemspec,swift,cpp,c," \
+          'java,py,clj,cljs,scala,js,yml,sh,json},{src,app,lib,test,' \
+          'spec,feature}/**/*.{rb,rake,gemspec,swift,cpp,c,java,py,clj,cljs,' \
+          "scala,js,yml,sh,json}}'"
         end
 
         def expect_punchlist_run(quality_checker)